Move oshi.ffm.driver.* to oshi.driver.* (mirror JNA driver location)
FFM driver classes are internal implementations (not exported via
JPMS), and per project convention should live alongside their JNA
equivalents in `oshi.driver.*` — same arrangement as the
`oshi.hardware.*` / `oshi.software.os.*` namespaces. The
`oshi.ffm.driver.*` location was an organic placement from the AIX
(#3350) and Solaris (#3339) FFM ports; the earlier Mac/Linux/Windows
FFM drivers already used `oshi.driver.*`, so this aligns the four
ports under one rulebook.

Moves (13 files + 2 package-info):

* `oshi.ffm.driver.unix.aix.PsInfoFFM` and the 7
  `oshi.ffm.driver.unix.aix.perfstat.Perfstat*FFM` drivers move to
  `oshi.driver.unix.aix.*`, next to the JNA `Perfstat*JNA` peers.
* `oshi.ffm.driver.unix.solaris.PsInfoFFM` and
  `oshi.ffm.driver.unix.solaris.kstat.SystemPagesFFM` move to
  `oshi.driver.unix.solaris.*`.
* `oshi.ffm.driver.windows.registry.InstalledAppsDataFFM` moves to
  `oshi.driver.windows.registry.*` next to the other Windows registry
  FFM drivers.

No JPMS export changes — `oshi.ffm.driver.*` was never exported, and
`oshi.driver.*` is also internal-only.

13 consumer files (in `oshi.hardware.platform.unix.{aix,solaris}.*`,
`oshi.software.os.unix.{aix,solaris}.*`, and
`oshi.software.os.windows.*`) had their imports rewritten by sed.
`module-info.test` add-opens lines updated to match the new package
names.

First commit in a multi-commit FFM package restructure; final state
ships as one PR. See [[project_ffm_structural_cleanup]] memory for
the full plan.
