+++ oshi-core-ffm/src/main/java/oshi/driver/unix/aix/perfstat/PerfstatMemoryFFM.java
@@ -2,7 +2,7 @@
  * Copyright 2026 The OSHI Project Contributors
  * SPDX-License-Identifier: MIT
  */
-package oshi.ffm.driver.unix.aix.perfstat;
+package oshi.driver.unix.aix.perfstat;
 
 import static oshi.ffm.unix.aix.PerfstatFunctions.PERFSTAT_MEMORY_TOTAL_T_SIZE;
 import static oshi.ffm.unix.aix.PerfstatFunctions.memPgspFree;
