+++ oshi-core-ffm/src/main/java/oshi/driver/unix/aix/perfstat/package-info.java
@@ -6,4 +6,4 @@
  * FFM-backed driver classes wrapping AIX {@code libperfstat} calls. Each class returns plain-Java mirror POJOs (inner
  * classes) for the fields OSHI consumes.
  */
-package oshi.ffm.driver.unix.aix.perfstat;
+package oshi.driver.unix.aix.perfstat;
