+++ oshi-core-ffm/src/main/java/oshi/hardware/platform/unix/solaris/SolarisGlobalMemoryFFM.java
@@ -10,7 +10,7 @@ import static oshi.util.Memoizer.memoize;
 import java.util.function.Supplier;
 
 import oshi.annotation.concurrent.ThreadSafe;
-import oshi.ffm.driver.unix.solaris.kstat.SystemPagesFFM;
+import oshi.driver.unix.solaris.kstat.SystemPagesFFM;
 import oshi.util.tuples.Pair;
 
 @ThreadSafe
