+++ oshi-core-ffm/src/main/java/oshi/software/os/unix/aix/AixOperatingSystemFFM.java
@@ -24,8 +24,8 @@ import java.util.stream.Collectors;
 import oshi.annotation.concurrent.ThreadSafe;
 import oshi.driver.common.unix.aix.Uptime;
 import oshi.driver.common.unix.aix.Who;
-import oshi.ffm.driver.unix.aix.perfstat.PerfstatConfigFFM;
-import oshi.ffm.driver.unix.aix.perfstat.PerfstatProcessFFM;
+import oshi.driver.unix.aix.perfstat.PerfstatConfigFFM;
+import oshi.driver.unix.aix.perfstat.PerfstatProcessFFM;
 import oshi.ffm.unix.aix.AixLibcFunctions;
 import oshi.software.common.AbstractOperatingSystem;
 import oshi.software.common.os.unix.aix.AixFileSystem;
