+++ oshi-core-ffm/src/main/java/oshi/software/os/windows/WindowsInstalledAppsFFM.java
@@ -6,7 +6,7 @@ package oshi.software.os.windows;
 
 import java.util.List;
 
-import oshi.ffm.driver.windows.registry.InstalledAppsDataFFM;
+import oshi.driver.windows.registry.InstalledAppsDataFFM;
 import oshi.software.os.ApplicationInfo;
 
 public final class WindowsInstalledAppsFFM {
