+++ oshi-core-ffm/src/test/java/module-info.test
@@ -15,9 +15,9 @@
 --add-opens
   com.github.oshi.ffm/oshi.ffm.unix.aix=org.junit.platform.commons
 --add-opens
-  com.github.oshi.ffm/oshi.ffm.driver.unix.aix=org.junit.platform.commons
+  com.github.oshi.ffm/oshi.driver.unix.aix=org.junit.platform.commons
 --add-opens
-  com.github.oshi.ffm/oshi.ffm.driver.unix.aix.perfstat=org.junit.platform.commons
+  com.github.oshi.ffm/oshi.driver.unix.aix.perfstat=org.junit.platform.commons
 --add-opens
   com.github.oshi.ffm/oshi.ffm.windows.com=org.junit.platform.commons
 --add-opens
